requirements.txt
asgiref==3.7.2
Django==5.0.3
django-cors-headers==4.3.1
django-debug-toolbar==4.4.2
djangorestframework==3.14.0
gunicorn==22.0.0
packaging==24.1
pytz==2024.1
sqlparse==0.4.4
